requirements.txt
alembic==1.5.1
aniso8601==8.1.0
atomicwrites==1.4.0
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
qcloudsms-py==0.1.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
wincertstore==0.2
zope.event==4.5.0
zope.interface==5.2.0
